fixed load function
removed ability to load a nonexistent level

+++ scripts/shortcuts.js
@@ -16,29 +16,36 @@ function paste(event) {
 function save(event) {
 	if (event.ctrlKey && event.key === 's') {
 		event.preventDefault();
-		localStorage.setItem('levelArray', JSON.stringify(levelArray));
-		localStorage.setItem('tilesPerRow', tilesPerRow);
-		localStorage.setItem('tilesPerColumn', tilesPerColumn);
-		localStorage.setItem('playStartX', playStartX);
-		localStorage.setItem('playStartY', playStartY);
-		localStorage.setItem('playEndX', playEndX);
-		localStorage.setItem('playEndY', playEndY);
-		console.log('saved the level');
+		if (confirm('Do you really want to save the level?\nThis will delete the previously saved level...')) {
+			localStorage.setItem('levelArray', JSON.stringify(levelArray));
+			localStorage.setItem('tilesPerRow', tilesPerRow);
+			localStorage.setItem('tilesPerColumn', tilesPerColumn);
+			localStorage.setItem('playStartX', playStartX);
+			localStorage.setItem('playStartY', playStartY);
+			localStorage.setItem('playEndX', playEndX);
+			localStorage.setItem('playEndY', playEndY);
+			console.log('saved the level');
+		}
 	}
 };
 
 function load(event) {
 	if (event.ctrlKey && event.key === 'l') {
-		event.preventDefault()
-		levelArray = JSON.parse(localStorage.getItem('levelArray'))
-		tilesPerRow = parseInt(localStorage.getItem('tilesPerRow'))
-		tilesPerColumn = parseInt(localStorage.getItem('tilesPerColumn'));
-		playStartX = parseInt(localStorage.getItem('playStartX'));
-		playStartY = parseInt(localStorage.getItem('playStartY'));
-		playEndX = parseInt(localStorage.getItem('playEndX'));
-		playEndY = parseInt(localStorage.getItem('playEndY'));
-		drawVisLevel()
-		console.log('loaded the level :3')
+		event.preventDefault();
+		if (localStorage.getItem('levelArray') === null) {
+			alert('cannot load a nonexistent level!!!!')
+		}
+		else {
+			levelArray = JSON.parse(localStorage.getItem('levelArray'))
+			tilesPerRow = parseInt(localStorage.getItem('tilesPerRow'))
+			tilesPerColumn = parseInt(localStorage.getItem('tilesPerColumn'));
+			playStartX = parseInt(localStorage.getItem('playStartX'));
+			playStartY = parseInt(localStorage.getItem('playStartY'));
+			playEndX = parseInt(localStorage.getItem('playEndX'));
+			playEndY = parseInt(localStorage.getItem('playEndY'));
+			drawVisLevel()
+			console.log('loaded the level :3')
+		}
 	} 
 }
